Added handling of user meta-data.
Manipulate a linked-list of key-value pairs, to create the meta-data list.
Then install it onto an iobuf.  This will be added to the object when it is
written.  During reading, a parser will retrieve these values and install
them onto the iobuf used for the get.  You can build a list of values once,
and reuse them many times,

Also, changed const-ness in the arguments of some functions.