Added support for streaming GET/PUT, and chunked-transfer-encoding.
User can provide a custom readfunc/writefunc.  This allows threaded
interaction with curl, such that a series of writes (or reads) can be
incrementally added to the data portion of a PUT (or GET).

There is example code in test_aws (cases 11 and 12).  These tests rely on
phtreads, which may not be available on all platforms that want to use
libaws4c.  Therefore, these tests are not compiled, by default.  If you
want to run them, you have to build with 'make ... PTHREADS=1'

Also added support for chunked-transfer-encoding.  This allows a PUT to be
sent when the total size of the final object is not known at the time the
PUT is invoked.  This could be combined with a streaming write.

Extended IOBuf so that it keeps track of how much unread data is available.
This can be useful for implementing a streaming readfunc.

Also extended IOBuf to provide a pointer to user data.  This can also be
useful in a threaded readfunc/writefunc.  In libaws4c, these functions
receive callbacks from curl, receiving a pointer to an IOBuf.  If they need
some other context, this context can be placed into IOBuf.user_data.

Finally, test_aws adds an error-message if loading of the user's
config-file fails.  Otherwise, this failure causes an obscure segfault at
runtime.