Moved IOBuf.emc_compatibility and CTE-flag to AWSContext.flags.
This fixes a problem where aws_iobuf_reset() was wiping out IOBuf.flags
where chunked-transfer-encoding was enabled.  The solution would be either
(a), add IOBuf.flags tot he things that are preserved across calls to
aws_iobuf_reset(), or (b) move it to context, which is already preserved.

TBD: In general, maybe a lot of the stuff that is currently preserved in
calls to aws_iobuf_reset() should just be moved into AWSContext, one way or
another.  Maybe it makes sense to keep IOBuf.user_data in IOBuf, and
preserve it through aws_iobuf_reset(), but many of the other things
(e.g. read_fn, write_fn, etc) should probably just be context-things.