aws_iobuf_reset() should not wipe everything then restore preserved values.
This behavior caused a subtle bug in MarFS, where the curl callback to a
custom readfunction happened at exactly the moment when the IOBuf.user_data
value was zeroed out, when another thread called aws_iobuf_reset().  The
other thread thought this was safe, because the custom readfunc always
waits on a semaphore before accessing the IOBuf contents. However, the
readfunc gets access to the semaphore via the user_data.  Even though
user_data was "not altered", in the long-run, by aws_iobuf_reset(), it was
actually being temporarily wiped, and then restored.

The upshot is that aws_iobuf_reset() can not temporarily wipe everything
and then restore selected values, unless we wrapp a lot more locking around
things.  The simpler approach is to tweak aws_iobuf_reset() so that it only
wipes those values that it is supposed to wipe, and leaves everything else
alone.