Allow GetStringToSign() to use a previously-computed time_t, for the date.
GetStringToSign() takes a new DateConv struct, instead of a char** to
receive the date that is computed internally.  If the time_t* inside
the DateConv is non-null, we use that, instead of the current time, to
generate the signature.

This allows a custom server to authenticate a signed request by
generating its own signature, using a date supplied in the request,
and a password looked up locally using a user-name supplied in a
request.